04_sql_and_python/01_challenge/date_transformation.py
import sqlite3
import pandas as pd

def transform_date_format(date_str):
    """
    Transforms a date string from 'YYYY-MM-DD' format to 'DD/MM/YYYY' format.
    If the input is not in the expected format, returns the original string.
    
    Args:
    date_str (str): A date string in 'YYYY-MM-DD' format.

    Returns:
    str: The transformed date string in 'DD/MM/YYYY' format or the original string if format is invalid.
    """
    # TODO: Implement the date format transformation logic
    pass

def apply_date_transformation(db_path, table_name, column_name):
    """
    Applies the transform_date_format function to a specified column in a SQLite table.

    Args:
    db_path (str): Path to the SQLite database file.
    table_name (str): Name of the table in the database.
    column_name (str): Name of the column to apply the date transformation.

    Returns:
    DataFrame: A pandas DataFrame with the transformed column data.
    """
    with sqlite3.connect(db_path) as conn:
        # TODO: Read the specified table into a DataFrame
        # TODO: Apply the transform_date_format function to the specified column
        # TODO: Return the modified DataFrame
        pass

# Example usage (To be modified by the learner)
if __name__ == "__main__":
    db_path = '../data/northwind.db'
    table_name = 'Orders' # Learner to confirm or change
    column_name = 'OrderDate' # Learner to confirm or change
    transformed_df = apply_date_transformation(db_path, table_name, column_name)
    print(transformed_df.head())
